refactor: dto instead cuz req and res
