+++ pkg/apperror/apperror.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/gofiber/fiber/v2/log"
-	"github.com/yokeTH/gofiber-template/pkg/response"
+	"github.com/yokeTH/gofiber-template/pkg/dto"
 )
 
 type AppError struct {
@@ -68,7 +68,7 @@ func ErrorHandler(c *fiber.Ctx, err error) error {
 	// if is app error
 	if IsAppError(err) {
 		e := err.(*AppError)
-		if err := c.Status(e.Code).JSON(response.ErrorResponse{Error: e.Message}); err != nil {
+		if err := c.Status(e.Code).JSON(dto.ErrorResponse{Error: e.Message}); err != nil {
 			// if can't send error
 			return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
 		}
@@ -81,7 +81,7 @@ func ErrorHandler(c *fiber.Ctx, err error) error {
 		code = e.Code
 	}
 
-	if err := c.Status(code).JSON(response.ErrorResponse{Error: err.Error()}); err != nil {
+	if err := c.Status(code).JSON(dto.ErrorResponse{Error: err.Error()}); err != nil {
 		// if can't send error
 		return c.Status(fiber.StatusInternalServerError).SendString("Internal Server Error")
 	}
