+++ pkg/dto/http.go
@@ -1,4 +1,4 @@
-package response
+package dto
 
 type ErrorResponse struct {
 	Error string `json:"error"`
